# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Mundo-3D API",
    "version": "1.0.0",
    "description": "REST API for the Mundo-3D storefront and admin backend."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "idUser": {
            "type": "integer"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "idRole": {
            "type": "integer",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          }
        },
        "required": ["idUser", "firstName", "lastName", "email", "image"]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": ["user"]
      },
      "UsersIndexResponse": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          }
        },
        "required": ["count", "users"]
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "m3d_auth",
        "description": "JWT issued on login/register, required by apiAuthMiddleware."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-csrf-token",
        "description": "Required by csrfGuard on unsafe methods (POST/PUT/PATCH/DELETE) for authenticated routes."
      }
    }
  },
  "paths": {
    "/users/login": {
      "post": {
        "summary": "Authenticate and receive session cookies",
        "description": "Sets m3d_auth (JWT, httpOnly), m3d_csrf and a display-data cookie. Also accepts `Email`/`Password` (normalized to lowercase before validation).",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "remember": {
                    "type": "boolean"
                  }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error."
          },
          "401": {
            "description": "Email or password do not match."
          },
          "429": {
            "description": "Rate-limited (loginLimiter)."
          }
        }
      }
    },
    "/users/register": {
      "post": {
        "summary": "Register a new user and receive session cookies",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 10
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 10
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 32,
                    "description": "Needs an uppercase, a digit and a special char."
                  },
                  "confirmPassword": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Required; .jpg/.png only."
                  }
                },
                "required": [
                  "firstName",
                  "lastName",
                  "email",
                  "password",
                  "confirmPassword",
                  "image"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "or user already exists.": null
          },
          "429": {
            "description": "Rate-limited (registerLimiter)."
          }
        }
      }
    },
    "/users/email-confirmation/confirm": {
      "post": {
        "summary": "Confirm an email address with an opaque token",
        "description": "Unauthenticated. The token is accepted only in the JSON request body.",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Confirmation completed or was already completed."
          },
          "400": {
            "description": "Invalid or expired confirmation token."
          }
        }
      }
    },
    "/users/email-confirmation/resend": {
      "post": {
        "summary": "Request an email confirmation resend",
        "description": "Unauthenticated. Syntactically valid requests receive the same acceptance response without disclosing account eligibility, send status, or rate-limit outcome. Local rate limiting is single-process only.",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted without disclosing account eligibility, send status, or rate-limit outcome.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": ["message"]
                }
              }
            }
          }
        }
      }
    },
...
    "/users/refresh": {
      "post": {
        "summary": "Exchange a valid refresh cookie for a fresh access token",
        "description": "Authenticates solely via the httpOnly, path-scoped m3d_refresh cookie — does NOT require a valid/unexpired m3d_auth. Not behind apiAuthMiddleware, exempt from csrfGuard, rate-limited (refreshLimiter). Rotates the presented token on every current use; a token superseded less than 30s ago (grace window) returns a fresh access cookie only and sets no refresh cookie (refresh-token-rotation spec).",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "Refreshed. Sets a new m3d_auth cookie, and a new m3d_refresh cookie only on rotation (never on a grace hit).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Refresh cookie absent",
            "expired": null,
            "revoked": null,
            "or replayed past its grace window.": null
          },
          "429": {
            "description": "Rate-limited (refreshLimiter)."
          }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "List all users (admin only)",
        "tags": ["Users"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "All users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersIndexResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Authenticated but not ADMIN."
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get the authenticated user's profile",
        "tags": ["Users"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The authenticated user's safe profile.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  },
                  "required": ["user"]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "404": {
            "description": "Authenticated principal no longer exists."
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get one user by id (admin only)",
        "tags": ["Users"],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated."
          },
          "403": {
            "description": "Authenticated but not ADMIN."
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
...
  },
...
}